requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2024.2.2
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
coverage==7.6.10
cycler==0.12.1
distro==1.9.0
docutils==0.21.1
et-xmlfile==1.1.0
Flask==3.1.0
flask-cors==5.0.1
Flask-MySQLdb==2.0.0
fonttools==4.51.0
git-filter-repo==2.47.0
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.8.2
kiwisolver==1.4.8
llvmlite==0.44.0
lxml==5.3.1
MarkupSafe==3.0.2
matplotlib==3.10.1
mutagen==1.47.0
mysql-connector-python==9.2.0
narwhals==1.31.0
numba==0.61.0
numpy==2.1.3
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.3
pillow==11.1.0
plotly==6.0.1
pluggy==1.5.0
pprintpp==0.4.0
pycountry==24.6.1
pycountry-convert==0.7.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.17.2
pyparsing==3.1.2
PySocks==1.7.1
pytest==8.3.5
pytest-cov==6.1.1
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
repoze.lru==0.7
requests==2.31.0
scipy==1.15.2
seaborn==0.13.2
selenium==4.30.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tk==0.1.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.1
waitress==3.0.2
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
